agent-api/requirements.txt
# Core FastAPI
fastapi[standard]>=0.113.0,<0.114.0
uvicorn>=0.22.0

# Pydantic for data validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# LangChain & integrations
langchain>=0.1.0
langchain-community>=0.0.20
langchain-ollama>=0.1.0

# Vector DB (for RAG)
chromadb>=0.4.0

# Document loaders (for PDFs, text, etc.)
pypdf>=3.0.0
unstructured>=0.11.0
python-multipart>=0.0.6

# Environment variables and utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Optional (for enhanced functionality)
sentence-transformers>=2.2.0
numpy>=1.24.0

# Testing
pytest>=7.0.0
httpx>=0.24.0